microservices/requirements.txt
fastapi[standard]
redis
datetime
polygon-api-client
pandas_market_calendars
dotenv
